move dummy page to pages dir
